Counterfactual Likelihood

Represents a counterfactual likelihood where one or more likelihood factors has been replaced with an intervention as specified by intervention_list. Inherits from Likelihood. Other factors (including their updates) are taken from an underlying observed_likelihood estimated from observed data.

CF_Likelihood

make_CF_Likelihood(...)

Arguments

...
Passes all arguments to the constructor. See documentation for the Constructor below.

Format

R6Class object.

Value

Likelihood object

Constructor

make_CF_Likelihood(observed_likelihood, intervention_list, ...)

observed_likelihood

Likelihood obect specifying the relevant factors of the observed likelihood

intervention_list

A list of objects inheriting from LF_base, representing the intervention.

...

Not currently used.

Fields

observed_likelihood

Likelihood obect specifying the relevant factors of the observed likelihood

intervention_list

A list of objects inheriting from LF_base, representing the intervention.

See also