Derived Likelihood Factor Estimated from Data + Other Likelihood values, using sl3.

Uses an sl3 learner to estimate a likelihood factor from data. Inherits from LF_base; see that page for documentation on likelihood factors in general.

Format

R6Class object.

Value

LF_base object

Constructor

define_lf(LF_fit, name, learner, ..., type = "density")

name

character, the name of the factor. Should match a node name in the nodes specified by tmle3_Task$npsem

learner

An sl3 learner to be used to estimate the factor

...

Not currently used.

type

character, either "density", for conditional density or, "mean" for conditional mean

Fields

learner

The learner or learner fit object

See also