Likelihood Factor Estimated from Data using NP-MLE

Uses NP-MLE (puts mass \(1/n\) on each of the observations, or uses weights if specified) to estimate a marginal density. Inherits from LF_base; see that page for documentation on likelihood factors in general. Only compatible with marginal likelihoods (no parent nodes). Only compatible with densities (no conditional means). The type argument will be ignored if specified.

LF_emp

Format

R6Class object.

Value

LF_base object

Constructor

define_lf(LF_emp, name, ...)

name

character, the name of the factor. Should match a node name in the nodes specified by tmle3_Task$npsem

...

Not currently used.

See also

Other Likelihood objects: CF_Likelihood, LF_base, LF_fit, LF_shift, LF_static, Likelihood, define_lf