Additive Effect of Treatment Among the Treated

Parameter definition for the Additive Effect of Treatment Among the Treated (ATT). Currently supports multiple static intervention nodes. Does yet not support dynamic rule or stochastic interventions.

Param_ATT

Format

R6Class object.

Value

Param_base object

Current Issues

clever covariates doesn't support updates; always uses initial (necessary for iterative TMLE, e.g. stochastic intervention)

doesn't integrate over possible counterfactuals (necessary for stochastic intervention)

clever covariate gets recalculated all the time (inefficient)

Constructor

define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)

observed_likelihood

A Likelihood corresponding to the observed likelihood

intervention_list_treatment

A list of objects inheriting from LF_base, representing the treatment intervention.

intervention_list_control

A list of objects inheriting from LF_base, representing the control intervention.

...

Not currently used.

outcome_node

character, the name of the node that should be treated as the outcome

Fields

cf_likelihood_treatment

the counterfactual likelihood for the treatment

cf_likelihood_control

the counterfactual likelihood for the control

intervention_list_treatment

A list of objects inheriting from LF_base, representing the treatment intervention

intervention_list_control

A list of objects inheriting from LF_base, representing the control intervention

See also

Other Parameters: Param_TSM, Param_base, Param_mean, define_param, tmle3_Fit