Base Class for Defining Parameters

A parameter is a function of the likelihood. Once given a Likelihood object, a parameter will a value. These objects also contain information about the efficient influence function (EIF) of a parameter, as well as its clever covariate(s).

Param_base

Format

R6Class object.

Value

Param_base object

Constructor

define_param(Param_base, observed_likelihood, ..., outcome_node)

observed_likelihood

A Likelihood corresponding to the observed likelihood

...

Not currently used.

outcome_node

character, the name of the node that should be treated as the outcome

Methods

clever_covariates(tmle_task = NULL)

Get the clever covariates for an TMLE update step.

tmle_task: tmle3_Task to get clever covariate values for. If NULL, the tmle_task used to train the observed likelihood will be used

estimates(tmle_task = NULL)

Get the parameter estimates and influence curve values.

tmle_task: tmle3_Task to get clever covariate values for. If NULL, the tmle_task used to train the observed likelihood will be used

Fields

observed_likelihood

the observed likelihood

outcome_node

character, the name of the outcome node

See also