TMLE fit object

A tmle_fit object, containing initial and updated estimates, as well as data about the fitting procedure. TMLE updates are calculated when the object is constructed.

tmle3_Fit

fit_tmle3(...)

Arguments

...
Passes all arguments to the constructor. See documentation for the Constructor.

Format

R6Class object.

Value

Param_base object

Constructor

fit_tmle3(tmle_task, likelihood, tmle_params, updater, max_it=100, ...)

tmle_task

A tmle3_Task object defining the data and NP-SEM

likelihood

A Likelihood object defining the factorized likelihood

tmle_params

A list of parameters inheriting from Param_base defining the parameter(s) of interest

updater

A tmle3_Update object defining the update procedure, including submodel and loss function

maxit

integer, maximum number of TMLE iterations

...

Not currently used.

Methods

set_timings(start_time, task_time, likelihood_time, params_time, fit_time)

Provide the timings for the different steps of the TMLE procedure, for later reporting to the user

tmle_task: tmle3_Task to get clever covariate values for. If NULL, the tmle_task used to train the observed likelihood will be used

estimates(tmle_task = NULL)

Get the parameter estimates and influence curve values.

tmle_task: tmle3_Task to get clever covariate values for. If NULL, the tmle_task used to train the observed likelihood will be used

Fields

tmle_task

A tmle3_Task object defining the data and NP-SEM

likelihood

A Likelihood object defining the factorized likelihood

tmle_params

A list of parameters inheriting from Param_base defining the parameter(s) of interest

tmle_names

A list of parameter names, obtained by calling param$name on each parameter

updater

A tmle3_Update object defining the update procedure, including submodel and loss function

steps

integer, he number of steps until TMLE converged

ED

vector, the mean of the EIF for all the parameters

initial_psi

vector, the initial parameter estimates

estimates

list, final parameter estimates and ICs

summary

data.table, summary of results

timings

data.frame, timings for each step (provided by tmle3_Fit$set_timings)

See also