A Node (set of variables) in an NPSEM

This class defines a node in an NPSEM

define_node(...)

Arguments

...
Passes all arguments to the constructor. See documentation for the Constructor below.

Format

R6Class object.

Value

tmle3_Node object

Constructor

make_tmle3_task(name, variables, parents = c(), variable_type = NULL)

name

character, the name of node

variables

character vector, the names of the variables that comprise the node

parents

character vector, the names of the parent nodes. If censoring, node is assumed to have no parents.

variable_type

variable_type object, specifying the data type of this variable. If censoring, variable_type will be guessed later from the data.

Methods

guess_variable_type(variable_data)

Guesses the variable_type from the provided data. This will be called by the tmle3_Task constructor if no variable_type was provided.

variable_data: the observed variable data.

Fields

name

character, the name of node

variables

character vector, the names of the variables that comprise the node

parents

character vector, the names of the parent nodes. If censoring, node is assumed to have no parents.

variable_type

variable_type object, specifying the data type of this variable.